MAINT: refactor handling and tests for division between object-dtype arrays and timedelta objects
